Change shell to use 2 threads.
Instead of using one single-threaded loop to handle user input and
execution, use two threads communicating with channels.

This is to prepare for change to support "keep-alive". Instead of trying
to spawn a new thread in keep_alive, instead grab a channel transmitter
from the shell state and then send a command to the execution thread.