Use KeepAliveTcpStream in ping.rs.
When the client sends out a keep alive ping, it also unnecessarily
resets the keep alive timeout. This should be okay for now?

If not, change ping to have a flag that will not reset the keep alive
timeout if present. And use that flag when keep alive timeout occurs.