Modify commands::connect() to wait for CONNACK.
Forgot to add this bit of functionality to connection process. The
client must send a connection packet and the broker needs to send a
CONNACK in response to successful connection attempt. The client must
close the connection if one is not received within a reasonable amount
of time.

Not sure what to do if the broker sends a packet other than CONNACK.
Currently the client will show an error and close the connection.