Make the keep_alive thread wait until connection established.
Add keep_alive suspend and resume so that pings are not sent out until
the Mqtt connection with broker is established. This may happen when you
spawn the tcp thread and then it takes > keep_alive time but <
CONNACK_TIMEOUT time for the broker to respond.

Change keep_alive::WakeReason to keep_alive::Msg, since the old name
doesn't really make sense and the new one is less verbose.