Add loop to shell to fix graphical cli most of the time.
This is a complex issue that is caused by having two threads running in
the shell run() function. And the reason for doing this is because stdin
blocks, but we need to be able to run commands while waiting on user
input.

Both threads are sychronized with two channels. The user presses enter,
and shell_cmd_tx is written to, unblocking the execution thread. The
input thread then blocks on rsp_queue_tx while the execution thread runs
the command. Upon finishing, it writes to rsp_queue_tx and blocks again on
shelL_cmd_tx.

What messes this up is that there are multiple produces that can write
into the shell_cmd_tx and this "desychronizes" the two threads. For
example, the keep_alive thread can write in an extra shell_cmd_tx ping
that generates an extra rsp_queue_tx packet. The "drain loop" on the
input thread fixes the graphical issues when this happens.

However, there is a corner case where the issue still happens. If the
execution thread pushes to the rsp_queue_tx right when the input thread
is between the drain loop and blocking on stdin, this still happens.
Luckily, this is resolved by pressing enter twice.

A proper fix would probbaly be to use an ncurses library to separate the user
input and console output into separate windows.