Implement enable set. This required a lot of changes...
Make it so that you can't call certain commands when it doesn't make
sense to do so. For example, you can't call disconnect if no connection
is present. It won't even show up in the help() function. That is what
is meant by "enable set".

Also important to note, that "shell" and "state" variables are mutex'd
between the shell threads. State AND THEN shell are locked in that
order.