Change mqtt-cli subscribe to send unsuback manually.
Trying to use the shell command queue results in a timeout, because the
command queue is not consumed until the current command is finished.

NOTE: That is definitely something to look into for the next iteration.
One can probably solve this issue by spawning a new thread to parse and
execute commands within the shell's execution thread. Or better yet, do
this using async lib instead.