Beef up Session struct and fix client code.
Give Session a proper constructor and make members private. Change all
accesses of session data to match.

Change BrokerMsg::ClientConnected to take a copy of Session instead of
the client id string. Maybe this should be a reference to Session in the
future? This might be complicated because it means we now need to give
BrokerMsg a lifetime...

Refactor out handle_client_disconnect into its own function for less
verbose code.