Remove TODO that is already done.
This is concerning the client sending packets before it receives a
connack. Due to the way the client handler is created, the packets will
always be handled in order and if the connection handshake fails, the
client will exit due to protocol violation because it will receive a non
connect packet while not in the Connected state.