Copy broker config into client session to get retry parameters.
According to stack overflow, the usual thing people do when they need to
keep references in struct members is that they just copy things instead
(if it's not too big). This avoids having to percolate a lifetime up
the entire struct hierarchy and figuring out how to bound lifetimes so
that mutable borrows aren't messed up (which looks impossible in this
case).