Implement wrapper type for mqttrs::Packet.
This type converts all mqttrs::Packet fields that are references into
owned types. The file also provides conversion functions to and from the
mqttrs::Packet type.

This is supposed to make it easier to code, having an mqtt::Packet type
that doesn't require lifetimes and associated buffers everywhere at the
cost of extra copying and memory fragmentation. The idea is that
eventually we can get rid of the mqttrs library and implement different
internals for the mqtt::Packet type and get some of the overhead from
copying back.